Changes integer format for vegas_pts column to decimal

+++ database/migrations/2016_09_02_144509_add_vegas_pts_column_to_game_lines_table.php
@@ -15,7 +15,7 @@ class AddVegasPtsColumnToGameLinesTable extends Migration
     {
         Schema::table('game_lines', function ($table) {
             
-            $table->integer('vegas_pts')->nullable()->after('pts');
+            $table->decimal('vegas_pts', 5, 2)->nullable()->after('pts');
         });
     }
